restructure docs and dm endpoints
add more docs and make it cleaner
