+++ sogs/routes/__init__.py
@@ -11,6 +11,7 @@ from .onion_request import onion_request as onion_request_endpoints
 from .rooms import rooms as rooms_endpoints
 from .messages import messages as messages_endpoints
 from .users import users as users_endpoints
+from .dm import dm as dm_endpoints
 
 from io import BytesIO
 
@@ -18,7 +19,7 @@ import qrencode
 
 from PIL.Image import NEAREST
 
-
+app.register_blueprint(dm_endpoints)
 app.register_blueprint(rooms_endpoints)
 app.register_blueprint(messages_endpoints)
 app.register_blueprint(users_endpoints)
