+++ sogs/routes/users.py
@@ -76,252 +76,6 @@ def extract_rooms_or_global(req, admin=True):
     return (None, True)
 
 
-def _serialize_message(msg):
-    return {
-        "id": msg.id,
-        "posted_at": msg.posted_at,
-        "expires_at": msg.expires_at,
-        "message": utils.encode_base64(msg.data),
-        "sender": msg.sender.session_id,
-        "recipient": msg.recipient.session_id,
-    }
-
-
-def _box(out: bool, *, since=None):
-    """ handle inbox/outbox endpoints common logic """
-    if not g.user.is_blinded:
-        abort(http.FORBIDDEN)
-    limit = utils.get_int_param('limit', 100, min=1, max=256, truncate=True)
-    get = Message.sent if out else Message.to
-    msgs = [_serialize_message(msg) for msg in get(user=g.user, limit=limit, since=since)]
-    if msgs or since is None:
-        return jsonify(msgs)
-    return Response('', status=http.NOT_MODIFIED)
-
-
-@users.get("/outbox")
-@auth.user_required
-def get_outbox():
-    """
-    Retrieves all of the messages we have sent to others via their inbox (up to `limit`).
-
-    # Query Parameters
-
-    The request takes an optional `limit` query parameter indicating the number of messages to
-    return (up to 256).  If omitted, at most 100 messages are returned.
-
-    # Return value
-
-    Returns a JSON array of up to `limit` (default 100) messages, with oldest messages first.  Each
-    element is a JSON object with keys:
-
-    - `id` — the unique integer message id.
-    - `posted_at` — unix timestamp (float) when the message was received by SOGS.
-    - `expires_at` — unix timestamp (float) when SOGS will expire and delete the message.
-    - `message` — the encrypted message body.
-    - `recipient` — the (blinded) Session ID of the recipient of the message.
-
-    # Encryption details
-
-    See [/inbox](#GET-inbox)
-
-    """
-    return _box(True)
-
-
-@users.get("/outbox/since/<int:msgid>")
-@auth.user_required
-def poll_outbox(msgid):
-    """
-    Polls for any DMs sent since the given id.
-
-    # URL Parameters
-
-    - `msgid` — the numeric message ID of the last sent message.  (Newer messages will always
-      have a higher ID).
-
-    # Query Parameters
-
-    The request takes an optional `limit` query parameter indicating the number of messages to
-    return (up to 256).  If omitted, at most 100 messages are returned.
-
-    # Return value
-
-    This method, on success, returns *either* a 200 (OK) status code with a list of 1 or more new
-    messages, or else returns a 304 (Not Modified) response with an empty body to indicate that
-    there are no messages for this user newer than the given ID.
-
-    If there are messages this returns a JSON array of up to `limit` (default 100) messages, with
-    oldest messages first.  Each element is exactly as described in the [all messages](#GET-outbox)
-    endpoint.
-    """
-    return _box(True, since=msgid)
-
-
-@users.get("/inbox")
-@auth.user_required
-def get_inbox():
-    """
-    Retrieves all of the user's current messages (up to `limit`).
-
-    # Query Parameters
-
-    The request takes an optional `limit` query parameter indicating the number of messages to
-    return (up to 256).  If omitted, at most 100 messages are returned.
-
-    # Return value
-
-    Returns a JSON array of up to `limit` (default 100) messages, with oldest messages first.  Each
-    element is a JSON object with keys:
-
-    - `id` — the unique integer message id.
-    - `posted_at` — unix timestamp (float) when the message was received by SOGS.
-    - `expires_at` — unix timestamp (float) when SOGS will expire and delete the message.
-    - `message` — the encrypted message body.
-    - `sender` — the (blinded) Session ID of the sender of the message.
-
-    # Encryption details
-
-    SOGS itself does not have the ability to decrypt the message contents and thus cannot enforce
-    any particular content; the following, however, is strongly recommended for Session client
-    interoperability:
-
-    Alice has master Session Ed25519 keypair a, A and blinded keypair ka, kA.
-
-    Bob has master Session Ed25519 keypair b, B and blinded keypair kb, kB.
-
-    Alice wants to send a message to Bob, knowing only kB (i.e. the blinded Session ID after
-    stripping the 0x15 prefix).
-
-    Alice constructs a message using Session protobuf encoding, then concatenates her *unblinded*
-    pubkey, A, to this message:
-
-    MSG || A
-
-    Alice then constructs an encryption key:
-
-    E = H(a * kB || kA || kB)
-
-    where H(.) is 32-byte BLAKE2b, and the `*` denotes unclamped Ed25519 scalar*point multiplication
-    (e.g. libsodium's `crypto_scalarmult_ed25519_noclamp`).
-
-    The `MSG || A` plaintext value is then encrypted using XChaCha20-Poly1305 (e.g. using
-    libsodium's `crypto_aead_xchacha20poly1305_ietf_encrypt` function), using a secure-random
-    24-byte nonce, no additional data, and encryption key `E`.
-
-    The final data message is then constructed as:
-
-    0x00 || CIPHERTEXT || NONCE
-
-    where 0x00 is a version byte (allowing for future alternative encryption formats) and the rest
-    are bytes.
-
-    Finally this is base64-encoded when send to/retrieved from the SOGS.
-
-    # Decryption
-
-    Decryption proceeds by reversing the steps above:
-
-    1. base64-decode the value.
-
-    2. Grab the version byte from the front and the 24-byte nonce from the back of the value.
-
-        a) if the version byte is not 0x00 abort because this message is from someone using a
-        different encryption protocol.
-
-    3. Construct the encryption key by calculating:
-
-            E = H(b * kA || kA || kB)
-
-        where kA is the sender's de-prefixed blinded Session ID, b is the user's master Ed25519 key,
-        and kB is the user's blinded Ed25519 for this SOGS server.
-
-    4. Decrypt the remaining ciphertext using the nonce.
-
-    5. Unpack the plaintext value into MSG and the sender's `A` value (i.e. the last 32 bytes).
-
-    6. Derive the sender's actual Session ID by converting `A` from an ed25519 pubkey to an
-       curve25519 pubkey and prepending 0x05.  (E.g. using libsodium's
-       `crypto_sign_ed25519_pk_to_curve25519` on `A`, then adding the 05 prefix to the front).
-
-    This then leaves the receiving client with the true Session ID of the sender, and the message
-    body (encoded according to typical Session message protobuf encoding).
-    """
-    return _box(False)
-
-
-@users.get("/inbox/since/<int:msgid>")
-@auth.user_required
-def poll_inbox(msgid):
-    """
-    Polls for any DMs received since the given id.
-
-    # URL Parameters
-
-    - `msgid` — the numeric message ID of the last received message.  (Newer messages will always
-      have a higher ID).
-
-    # Query Parameters
-
-    The request takes an optional `limit` query parameter indicating the number of messages to
-    return (up to 256).  If omitted, at most 100 messages are returned.
-
-    # Return value
-
-    This method, on success, returns *either* a 200 (OK) status code with a list of 1 or more new
-    messages, or else returns a 304 (Not Modified) response with an empty body to indicate that
-    there are no messages for this user newer than the given ID.
-
-    If there are messages this returns a JSON array of up to `limit` (default 100) messages, with
-    oldest messages first.  Each element is exactly as described in the [all messages](#GET-inbox)
-    endpoint.
-    """
-    return _box(False, since=msgid)
-
-
-@users.post("/inbox/<BlindSessionID:sid>")
-@auth.user_required
-def send_inbox(sid):
-    """
-    Delivers a direct message to a user via their blinded Session ID.
-
-    The body of this request is a JSON object containing a `message` key with a value of the
-    encrypted-then-base64-encoded message to deliver.
-
-    Message encryption is described in the [`GET` /inbox](#GET-inbox) endpoint.
-
-    # Return value
-
-    On successful deposit of the message a 201 (Created) status code is returned.  The body will be
-    a JSON object containing an `expires_at` key indicating the unix timestamp of when the message
-    expires.
-
-    # Error status codes
-
-    400 Bad Request — if no message is provided.
-
-    404 Not Found — if the given Session ID does not exist on this server, either because they have
-    never accessed the server, or because they have been permanently banned.
-    """
-    try:
-        recip_user = User(session_id=sid, autovivify=False)
-    except NoSuchUser:
-        abort(http.NOT_FOUND)
-
-    if recip_user.banned:
-        abort(http.NOT_FOUND)
-
-    req = request.json
-    message = req.get('message')
-    if message is None:
-        app.logger.warning("No message provided")
-        abort(http.BAD_REQUEST)
-
-    with db.transaction():
-        msg = Message(data=utils.decode_base64(message), recip=recip_user, sender=g.user)
-    return jsonify({"expires_at": msg.expires_at}), http.CREATED
-
-
 @users.post("/user/<SessionID:sid>/moderator")
 @auth.user_required
 def set_mod(sid):
